+++ web/models/image.ex
@@ -40,7 +40,8 @@ defmodule Brando.Image do
   """
   @spec changeset(t, atom, Keyword.t | Options.t) :: t
   def changeset(model, :create, params) do
-    cast(model, params, @required_fields ++ @optional_fields)
+    model
+    |> cast(params, @required_fields ++ @optional_fields)
     |> validate_required(@required_fields)
   end