+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v0.26.1 (2016-03-28)
+
+* Enhancements
+  * Add terminal notification at the end of fabric release build.
+
+* Bug fixes
+  * Fix :textarea not being able to call default function
+
 ## v0.26.0 (2016-03-22)
 
 * Enhancements